Mega update: restructure project, add libDaisy and DaisySP as submodules, add build scripts, add VCA firmware, provide pre-compiled firmware in bin directory, improve README to reflect changes.
