+++ src/smoodge-delay/Makefile
@@ -1,0 +1,13 @@
+# Project Name
+TARGET = smoodge-delay
+
+# Sources
+CPP_SOURCES = smoodge-delay.cpp
+
+# Library Locations
+LIBDAISY_DIR = ../../lib/libDaisy
+DAISYSP_DIR = ../../lib/DaisySP
+
+# Core location, and generic Makefile.
+SYSTEM_FILES_DIR = $(LIBDAISY_DIR)/core
+include $(SYSTEM_FILES_DIR)/Makefile
