+++ src/vca/Makefile
@@ -1,0 +1,15 @@
+# Project Name
+TARGET = vca
+
+USE_DAISYSP_LGPL=1
+
+# Sources
+CPP_SOURCES = vca.cpp
+
+# Library Locations
+LIBDAISY_DIR = ../../lib/libDaisy
+DAISYSP_DIR = ../../lib/DaisySP
+
+# Core location, and generic Makefile.
+SYSTEM_FILES_DIR = $(LIBDAISY_DIR)/core
+include $(SYSTEM_FILES_DIR)/Makefile
