+++ src/vca/README.md
@@ -1,0 +1,14 @@
+# VCA
+
+## Author
+
+Ben van der Burgh
+
+## Description
+
+Lets you use the Löwenzahnhonig as a dual VCA.
+
+ - Pot 1: left offset
+ - Pot 2: right offset
+ - Pot 3: left CV attenuation
+ - Pot 4: right CV attenuation
